+++ game/race/helpers.js
@@ -1,24 +1,6 @@
 const User = require("../../models/User");
 const GOLDPRIZE = 500;
-
-// https://medium.com/javascript-in-plain-english/how-to-deep-copy-objects-and-arrays-in-javascript-7c911359b089
-const deepCopyFunction = (inObject) => {
-    let value, key;
-
-    if (typeof inObject !== "object" || inObject === null) {
-      return inObject;
-    }
-
-    const outObject = Array.isArray(inObject) ? [] : {};
-
-    for (key in inObject) {
-      value = inObject[key];
-
-      outObject[key] = deepCopyFunction(value);
-    }
-
-    return outObject;
-  };
+const { asyncForEach } = require("../_GLOBAL_HELPERS");
 
 const createChanceArray = (raceDataCopy)=> {
     const result = [];
@@ -53,17 +35,10 @@ const racePayOut = async (event) =>{
     return;
 };
 
-async function asyncForEach(array, callback) {
-    for (let index = 0; index < array.length; index += 1) {
-      await callback(array[index], index, array);
-    }
-  }
-
-
     const validateUser = async (discordId) =>{
         const user = await User.findOne({ "account.userId": discordId }).lean();
         return user.resources.gold >= GOLDPRIZE;
     };
 
 
-module.exports = { createChanceArray, racePayOut, asyncForEach, validateUser, deepCopyFunction };
+module.exports = { createChanceArray, racePayOut, validateUser };
